adding the stop word function that removes stop words from the list. Fixed an error in the hm_remove function where the word being removed was the only one in the bucket

+++ Project2/hashmap.h
@@ -5,7 +5,7 @@ struct llnode {
         char* word;
         //char* document_id;
         //int num_occurrences;
-        double df_score;
+        int df_score;
         struct llnode* next;
         struct lldoc* docptr;
 };
